Use pl.DataFrame for Arrow stream conversion; Polars 2.0 compatibility
- read_sav/scan_sav: convert PyCapsule streams with pl.DataFrame() instead of
  pl.from_arrow(). Polars >= 1.44 emits a FutureWarning for from_arrow on
  stream objects, and Polars 2.0 returns a Series there, which broke both
  readers. The constructor path has been available since Polars 1.3.
- codebook type detection: cast to Float64 before is_in() so Int64 columns
  work under Polars 2.0's strict is_in coercion.
- tests: build Date test column from date objects (String->Date cast removed
  in Polars 2.0); add tests/test_polars_compat.py guarding the conversion
  paths with FutureWarning treated as error; add it to CI.

+++ python/ambers/codebook/_types.py
@@ -355,7 +355,7 @@ def detect_types(
     binary_info: dict[str, str] = {}
     if cols_to_check:
         exprs = (
-            [pl.col(c).drop_nulls().is_in([0.0, 1.0]).all().alias(f"{c}::bin")
+            [pl.col(c).drop_nulls().cast(pl.Float64).is_in([0.0, 1.0]).all().alias(f"{c}::bin")
              for c in cols_to_check]
             + [pl.col(c).drop_nulls().n_unique().alias(f"{c}::nu")
                for c in cols_to_check]
